pyproject.toml
[tool.poetry]
name = "pythonanywhere-api"
version = "0.1.0"
description = "PythonAnyWhere API cli tool"
authors = ["Kirill Lekhov"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7"
python-dotenv = "0.21.1"
requests = "^2.31.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
